% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessAPI.R
\name{baato_API}
\alias{baato_API}
\title{initialize key}
\description{
initialize key

initialize key
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-baato_API-new}{\code{baato_API$new()}}
\item \href{#method-baato_API-set_key}{\code{baato_API$set_key()}}
\item \href{#method-baato_API-display_key}{\code{baato_API$display_key()}}
\item \href{#method-baato_API-validate_resp}{\code{baato_API$validate_resp()}}
\item \href{#method-baato_API-display_resp}{\code{baato_API$display_resp()}}
\item \href{#method-baato_API-rev_api_call}{\code{baato_API$rev_api_call()}}
\item \href{#method-baato_API-search_api_call}{\code{baato_API$search_api_call()}}
\item \href{#method-baato_API-places_api_call}{\code{baato_API$places_api_call()}}
\item \href{#method-baato_API-nearby_api_call}{\code{baato_API$nearby_api_call()}}
\item \href{#method-baato_API-styles_api_call}{\code{baato_API$styles_api_call()}}
\item \href{#method-baato_API-baato_directions_api}{\code{baato_API$baato_directions_api()}}
\item \href{#method-baato_API-clone}{\code{baato_API$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-new"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$new(key = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{set key}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-set_key"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-set_key}{}}}
\subsection{Method \code{set_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$set_key(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{display key}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-display_key"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-display_key}{}}}
\subsection{Method \code{display_key()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$display_key()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{validate the response 
Check if the json object returned by API call is valid}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-validate_resp"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-validate_resp}{}}}
\subsection{Method \code{validate_resp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$validate_resp(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
error message 
display the response from the API request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-display_resp"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-display_resp}{}}}
\subsection{Method \code{display_resp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$display_resp(req)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{req}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format 
Calls the reverse api
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-rev_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-rev_api_call}{}}}
\subsection{Method \code{rev_api_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$rev_api_call(lat, lon)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{}

\item{\code{lon}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-search_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-search_api_call}{}}}
\subsection{Method \code{search_api_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$search_api_call(q, limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{}

\item{\code{limit}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format 
Calls the places api
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-places_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-places_api_call}{}}}
\subsection{Method \code{places_api_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$places_api_call(placeId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{placeId}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format 
Calls nearby api
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-nearby_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-nearby_api_call}{}}}
\subsection{Method \code{nearby_api_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$nearby_api_call(type, lat, lon, limit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{}

\item{\code{lat}}{}

\item{\code{lon}}{}

\item{\code{limit}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format 
Calls the styles api
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-styles_api_call"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-styles_api_call}{}}}
\subsection{Method \code{styles_api_call()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$styles_api_call(style_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style_name}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format 
Calls the directions api
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-baato_directions_api"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-baato_directions_api}{}}}
\subsection{Method \code{baato_directions_api()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$baato_directions_api(points, mode)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{points}}{}

\item{\code{mode}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the API response in JSON format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-baato_API-clone"></a>}}
\if{latex}{\out{\hypertarget{method-baato_API-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{baato_API$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
