% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFarms.R
\name{getFarms}
\alias{getFarms}
\title{this function returns a data.frame with all farms in a given group}
\usage{
getFarms(group = "", token = "")
}
\arguments{
\item{group}{a string id of the group, for intance "GR-96c6c34e-e884-483e-94af-6fabc055c4bd"}

\item{token}{a string with the token from getToken function}
}
\value{
a data frame of all farms in the group
}
\description{
this function returns a data.frame with all farms in a given group
}
