% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroups.R
\name{getGroups}
\alias{getGroups}
\title{this function returns a data.frame with all groups accessible with a given token}
\usage{
getGroups(token = "")
}
\arguments{
\item{token}{a string with the token from getToken function}
}
\value{
a data frame of all groups available with the token
}
\description{
this function returns a data.frame with all groups accessible with a given token
}
