% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObs.R
\name{getObs}
\alias{getObs}
\title{get observations in a group
get all observations between 2 dates made in a group
return a list with 2 elements : a data frame of observations and the name of cols}
\usage{
getObs(
  group = "",
  token = "",
  startDate = "2022-01-01",
  endDate = "2022-12-31"
)
}
\arguments{
\item{group}{a string id of the group, for intance "GR-96c6c34e-e884-483e-94af-6fabc055c4bd"}

\item{token}{a string with the token from getToken function}

\item{startDate}{a string with the start date in format YYYY-MM-DD}

\item{endDate}{a string with the end date in format YYYY-MM-DD}
}
\description{
get observations in a group
get all observations between 2 dates made in a group
return a list with 2 elements : a data frame of observations and the name of cols
}
