% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlots.R
\name{getPlots}
\alias{getPlots}
\title{this function returns a data.frame with all parcels in a group}
\usage{
getPlots(group = "", token = "")
}
\arguments{
\item{group}{a string id of the group, for intance "GR-96c6c34e-e884-483e-94af-6fabc055c4bd"}

\item{token}{a string with the token from getToken function}
}
\value{
a data frame of all groups available with the token
}
\description{
this function returns a data.frame with all parcels in a group
}
