% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{this function returns a token after authentication to landfiles API}
\usage{
getToken(
  url = "https://api.landfiles.fr/api/authenticationservice/auth/oauth/token",
  username = "",
  password = ""
)
}
\arguments{
\item{url}{is a string with the url of the API, by default https://api.landfiles.fr/api/authenticationservice/auth/oauth/token}

\item{username}{is a string}

\item{password}{is a string}
}
\value{
if the request is succesful the function returns a string token
}
\description{
this function returns a token after authentication to landfiles API
}
